Fix compiler warnings
reg. size_t, some other comparisons between signed and unsigned and some
unhandled cases. Also some unused variables and paremeters, a possibly
unintilalized variable and a possible return without value.

+++ src/engraving/playback/renderers/gracenotesrenderer.h
@@ -59,7 +59,7 @@ private:
                                          const mpe::duration_t duration, const mpe::timestamp_t timestamp, mpe::PlaybackEventList& result);
 
     static mpe::duration_t graceNotesMaxAvailableDuration(const mpe::ArticulationType type, const RenderingContext& ctx,
-                                                          const int graceNotesCount);
+                                                          const size_t graceNotesCount);
     static mpe::timestamp_t graceNotesStartTimestamp(const mpe::ArticulationType type, const mpe::duration_t availableDuration,
                                                      const mpe::timestamp_t& nominalTimestamp);
